# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Novus Markets Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /markets:
    get:
      operationId: getMarkets
      tags: [markets]
      summary: Get available markets
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [forex, crypto, indices, commodities]
      responses:
        "200":
          description: Markets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Market"
...
  /markets/candles:
    get:
      operationId: getCandles
      tags: [markets]
      summary: Get candlestick data for a symbol
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: timeframe
          in: query
          schema:
            type: string
            enum: [1m, 5m, 15m, 1h, 4h, 1D]
            default: 1h
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Candle data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candle"

  # ── Trades ──────────────────────────────────────────────────────────────────
  /trades:
    get:
      operationId: getTrades
      tags: [trades]
      summary: Get trades for an account
      parameters:
        - name: accountId
          in: query
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [open, pending, closed]
      responses:
        "200":
          description: Trades list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
    post:
      operationId: executeTrade
      tags: [trades]
      summary: Execute a trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Trade executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/{id}/close:
    post:
      operationId: closeTrade
      tags: [trades]
      summary: Close an open trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/summary:
    get:
      operationId: getTradeSummary
      tags: [trades]
      summary: Get trade performance summary
      parameters:
        - name: accountId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeSummary"

  # ── Deposits & Withdrawals ────────────────────────────────────────────────
  /deposits:
    get:
      operationId: getDeposits
      tags: [deposits]
      summary: Get deposit/withdrawal history
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [deposit, withdrawal]
      responses:
        "200":
          description: Transaction history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createDeposit
      tags: [deposits]
      summary: Create a deposit request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Deposit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /deposits/withdraw:
    post:
      operationId: createWithdrawal
      tags: [deposits]
      summary: Create a withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /deposits/fee-preview:
    get:
      operationId: getDepositFeePreview
      tags: [deposits]
      summary: Preview deposit fee breakdown
      parameters:
        - name: amount
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Fee breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeePreview"

  # ── Notifications ─────────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Market:
      type: object
      required: [id, symbol, name, category, bidPrice, askPrice, spread, change24h, volume24h]
      properties:
        id:
          type: integer
        symbol:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [forex, crypto, indices, commodities]
        bidPrice:
          type: number
        askPrice:
          type: number
        spread:
          type: number
        change24h:
          type: number
        volume24h:
          type: number
        high24h:
          type: number
        low24h:
          type: number
        pipSize:
          type: number
...
    Candle:
      type: object
      required: [time, open, high, low, close, volume]
      properties:
        time:
          type: integer
          description: Unix timestamp in seconds
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number

    # ── Trades ─────────────────────────────────────────────────────────────
    Trade:
      type: object
      required: [id, accountId, symbol, side, type, lotSize, openPrice, status, createdAt]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        type:
          type: string
          enum: [market, limit, stop]
        lotSize:
          type: number
        openPrice:
          type: number
        closePrice:
          type: ["number", "null"]
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        leverage:
          type: integer
        margin:
          type: number
        pnl:
          type: ["number", "null"]
        fee:
          type: number
        status:
          type: string
          enum: [open, pending, closed, cancelled]
        createdAt:
          type: string
          format: date-time
        closedAt:
          type: ["string", "null"]
    TradeInput:
      type: object
      required: [accountId, symbol, side, type, lotSize, leverage]
      properties:
        accountId:
          type: integer
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        type:
          type: string
          enum: [market, limit, stop]
        lotSize:
          type: number
        leverage:
          type: integer
        limitPrice:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
    TradeSummary:
      type: object
      required: [totalTrades, winRate, totalPnl, totalFees, avgTradeSize]
      properties:
        totalTrades:
          type: integer
        openTrades:
          type: integer
        winRate:
          type: number
        totalPnl:
          type: number
        totalFees:
          type: number
        avgTradeSize:
          type: number
        bestTrade:
          type: ["number", "null"]
        worstTrade:
          type: ["number", "null"]

    # ── Deposits ───────────────────────────────────────────────────────────
    Transaction:
      type: object
      required: [id, userId, type, amount, fee, netAmount, method, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        accountId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [deposit, withdrawal]
        amount:
          type: number
        fee:
          type: number
        netAmount:
          type: number
        method:
          type: string
          enum: [mpesa, crypto_usdt, crypto_btc, card, bank_transfer]
        status:
          type: string
          enum: [pending, approved, rejected, completed]
        reference:
          type: ["string", "null"]
        walletAddress:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        reviewedAt:
          type: ["string", "null"]
    DepositInput:
      type: object
      required: [amount, method, accountId]
      properties:
        accountId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [mpesa, crypto_usdt, crypto_btc, card, bank_transfer]
        phoneNumber:
          type: string
        walletAddress:
          type: string
        cardLast4:
          type: string
    WithdrawalInput:
      type: object
      required: [amount, method, accountId]
      properties:
        accountId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [mpesa, crypto_usdt, crypto_btc, card, bank_transfer]
        phoneNumber:
          type: string
        walletAddress:
          type: string
    FeePreview:
      type: object
      required: [depositAmount, feeRate, feeAmount, netAmount]
      properties:
        depositAmount:
          type: number
        feeRate:
          type: number
        feeAmount:
          type: number
        netAmount:
          type: number

    # ── Notifications ─────────────────────────────────────────────────────
    Notification:
      type: object
      required: [id, userId, title, message, type, read, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [trade, deposit, kyc, system, alert]
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ── Settings ───────────────────────────────────────────────────────────
...